Abstract useEffect and states into new hook
